+++ .env.example
@@ -3,3 +3,4 @@ LAST_FM_API_SECRET=your_last_fm_api_secret
 LASTFM_SESSION=your_last_fm_session_key_will_be_auto_generated
 YTMUSIC_AUTH_KEY=your_generated_encryption_key
 DISCORD_WEBHOOK_URL=your_discord_webhook_url_optional
+SCROBBLE_TIMEZONE=Asia/Kolkata
