+++ GITHUB_ACTIONS_GUIDE.md
@@ -24,6 +24,10 @@ You need to add the following secrets to your GitHub repository. It is recommend
 -   `YTMUSIC_AUTH_KEY`: The AES key generated by `encrypt_auth.py` for `browser.json.enc`
 -   `DISCORD_WEBHOOK_URL`: Your Discord webhook URL for failure/success notifications
 
+### Optional Environment Variable (Non-Secret)
+
+-   `SCROBBLE_TIMEZONE`: IANA timezone used for report date/listening flow buckets (default: `Asia/Kolkata`)
+
 ### Security: Encrypting your YouTube Music Auth
 
 Instead of using raw cookies or committing `browser.json`, we use AES encryption:
@@ -145,6 +149,7 @@ jobs:
 ### How it Works
 
 - **`on.schedule`**: Runs once daily at 23:55 IST (18:25 UTC)
+- **Timezone note**: GitHub cron is UTC. If you change local timezone but want to keep 23:55 local execution, update the cron UTC conversion in `.github/workflows/sync.yml`.
 - **`on.push`**: Runs when code is pushed to master branch
 - **`on.workflow_dispatch`**: Allows manual triggering
 - **Cookie validation**: The script validates your YouTube Music cookie before processing
